// src/app/core/enum/structure.enum.ts
export enum Structure {
    BOARD = 0,
    DIRECTOR = 1,
    VICE_DIRECTOR = 2,
    DEPARTMENT_MANAGER = 3,
    WORKER = 4
  }

  // export enum Structure {
  //   BOARD = "zarząd",
  //   DIRECTOR = "dyrektor",
  //   VICE_DIRECTOR = "zastępca dyrektora",
  //   DEPARTMENT_MANAGER = "kierownik działu",
  //   WORKER = "pracownik"
  // }